+++ etsi_its_msgs_utils/include/etsi_its_msgs_utils/impl/cam/cam_getters.h
