+++ etsi_its_msgs_utils/include/etsi_its_msgs_utils/impl/cam/cam_utils.h
@@ -29,8 +29,8 @@ SOFTWARE.
  * @brief Utility functions for the ETSI ITS CAM
  */
 
-#include <etsi_its_msgs/impl/cdd/cdd_checks.h>
-#include <etsi_its_msgs/impl/constants.h>
+#include <etsi_its_msgs_utils/impl/cdd/cdd_checks.h>
+#include <etsi_its_msgs_utils/impl/constants.h>
 
 #pragma once
