+++ etsi_its_msgs_utils/include/etsi_its_msgs_utils/impl/cdd/cdd_getters.h
