+++ etsi_its_msgs_utils/include/etsi_its_msgs_utils/impl/constants.h
