+++ etsi_its_msgs_utils/test/test_cam_access.ros2.cpp
@@ -1,7 +1,7 @@
 
 #include <etsi_its_cam_msgs/msg/cam.hpp>
 
-#include <etsi_its_msgs/cam_access.hpp>
+#include <etsi_its_msgs_utils/cam_access.hpp>
 
 using namespace etsi_its_cam_msgs::msg;
 using namespace etsi_its_cam_msgs::access;
