+++ etsi_its_rviz_plugins/include/displays/CAM/cam_render_object.hpp
@@ -29,7 +29,7 @@ SOFTWARE.
 #include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 
 #include <tf2/LinearMath/Quaternion.h>
-#include <etsi_its_msgs/cam_access.hpp>
+#include <etsi_its_msgs_utils/cam_access.hpp>
 
 #include "rviz_common/validate_floats.hpp"
